% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam.R
\name{adam}
\alias{adam}
\title{Read ADaM datasets from Pentara's file structure}
\usage{
adam(name)
}
\arguments{
\item{name}{The name of a SAS dataset in the ADaM directory.}
}
\value{
invisibly returns NULL
}
\description{
Automatically fill in file paths based on Pentara SOP file
structures to make reading ADaM datasets into scripts easier.
}
\examples{
\dontrun{adam("adsl")}

}
