% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pentara-package.R
\docType{package}
\name{pentara-package}
\alias{pentara}
\alias{pentara-package}
\title{pentara: Simplify Common Pentara Workflows}
\description{
Provide tools to simplify common work flows for R users at Pentara Corporation.
}
\keyword{internal}
