% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sforest.R
\name{sforest}
\alias{sforest}
\title{Classic forest plot}
\usage{
sforest(
  dat,
  res,
  ci.lb,
  ci.ub,
  ordervar,
  orderlab,
  suborder,
  sublab,
  xlab,
  footnote = "",
  varlist = NULL,
  varlabs = NULL,
  pvals = NULL,
  refline = 0,
  col = "aliceblue",
  wd1 = 0.5,
  wd2 = 0.5,
  alim = NULL,
  col1 = " ",
  bold.col = "black",
  latex.out = TRUE,
  express = FALSE,
  expresso = FALSE,
  highbold = F,
  digits = 2,
  refcol = "black",
  show = NULL,
  nobold = FALSE,
  wgts = NULL,
  trendthresh = 0.05,
  trend.col = "black",
  at.lab = NULL,
  ci.lab = NULL,
  na.action = NULL,
  pdig = 4,
  graph.lab = "Graphical Summary",
  ci.correct = 0,
  scale.override = NULL,
  ...
)
}
\arguments{
\item{dat}{data for the plot}

\item{res}{a string naming a column to plot as points}

\item{ci.lb}{a string naming column to plot as ci upper bounds}

\item{ci.ub}{a string naming column to plot as ci lower bounds}

\item{ordervar}{a string naming a column that defines the order of sections. The column must be numeric and have the same value for all rows in the same section.}

\item{orderlab}{a string naming a column of section labels. The column must be character and have the same value for all rows in the same section.}

\item{suborder}{a string naming a column that defines the order within sections. The column must be numeric and have different values for all rows in the same section.}

\item{sublab}{a string naming a column of row labels. The column must be character and have different values for all rows in the same section.}

\item{xlab}{a string to be used as the x-axis label}

\item{varlist}{a character vector containing the names of other columns to be included in the plot}

\item{varlabs}{a character vector containing the labels to be used in the plot for variables in varlist}

\item{pvals}{a character vector containing the names of columns in varlist that should be formatted as p-values}

\item{...}{}
}
\value{
invisibly returns NULL
}
\description{
Creates forest plot that matches Pentara's historic forest plot style.
}
\examples{
forest_dat <- data.frame(
  mean = c(.1, .5, .2, .8),
  lower = c(.1, .5, .2, .8) - qnorm(.975),
  upper = c(.1, .5, .2, .8) + qnorm(.975),
  ordervar = rep(c(1,2), each = 2),
  orderlab = rep(c("Study 1", "Study 2"), each = 2),
  subordervar = rep(c(1,2), times = 2),
  suborderlab = rep(c("Endpoint 1", "Endpoint 2"), times = 2),
  p_val = c(.3756, .1023, .1542, .043)
)

sforest(
  dat = forest_dat,
  res = "mean",
  ci.lb = "lower",
  ci.ub = "upper",
  ordervar = "ordervar",
  orderlab = "orderlab",
  suborder = "subordervar",
  sublab = "suborderlab",
  varlist = c("p_val"),
  varlabs = c("P-Value"),
  pvals = c("p_val"),
  xlab = "Treatment Effect",
  latex.out = FALSE
)
}
