% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippets.R
\name{snippets}
\alias{snippets}
\title{Import package snippets}
\usage{
snippets()
}
\value{
boolean invisible(FALSE) if nothing was added, invisible(TRUE) if snippet definitions were added
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
\code{snippets} copies all (missing) snippet definitions
in \code{inst/rstudio/Rsnippets.txt} to the RStudio user snippet file.
}
\examples{
\dontrun{snippets()}
}
