% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippets.R
\name{snippets}
\alias{snippets}
\title{Export snippets}
\usage{
snippets()
}
\value{
boolean invisible(FALSE) if nothing was added, invisible(TRUE) if snippet definitions were added
}
\description{
\code{snippets} copies all (missing) snippet definitions
  in 'inst/rstudio/Rsnippets.txt' and 'Rmdsnippets.txt' to the RStudios user snippet location.
}
\examples{
\dontrun{snippets()}
}
